requirements.txt
ffprobe-python
ffmpeg-python
discord.py
pyjson5
psutil
Pillow
yt-dlp
pydub